+++ Sources/VComponents/Components/Inputs/TextField/VTextField.swift
@@ -176,7 +176,7 @@ public struct VTextField: View {
     @ViewBuilder private var header: some View {
         if let headerTitle, !headerTitle.isEmpty {
             VText(
-                type: .multiLine(alignment: .leading, lineLimit: uiModel.layout.headerLineLimit),
+                type: uiModel.layout.headerTitleType,
                 color: uiModel.colors.header.value(for: internalState),
                 font: uiModel.fonts.header,
                 text: headerTitle
@@ -188,7 +188,7 @@ public struct VTextField: View {
     @ViewBuilder private var footer: some View {
         if let footerTitle, !footerTitle.isEmpty {
             VText(
-                type: .multiLine(alignment: .leading, lineLimit: uiModel.layout.footerLineLimit),
+                type: uiModel.layout.footerTitleType,
                 color: uiModel.colors.footer.value(for: internalState),
                 font: uiModel.fonts.footer,
                 text: footerTitle