+++ Sources/VComponents/Components/Inputs/TextField/VTextFieldUIModel.swift
@@ -72,11 +72,11 @@ public struct VTextFieldUIModel {
         /// Spacing between text and buttons. Defaults to `10`.
         public var contentSpacing: CGFloat = 10
         
-        /// Header line limit. Defaults to `1`.
-        public var headerLineLimit: Int? = 1
+        /// Header title type. Defaults to `singleline`.
+        public var headerTitleType: VTextType = .singleLine
         
-        /// Footer line limit. Defaults to `5`.
-        public var footerLineLimit: Int? = 5
+        /// Footer title type. Defaults to `multiline` of `1...5` lines.
+        public var footerTitleType: VTextType = .multiLine(alignment: .leading, lineLimit: 1...5)
 
         /// Spacing between header, textfield, and footer. Defaults to `3`.
         public var headerTextFieldFooterSpacing: CGFloat = segmentedPickerReference.layout.headerPickerFooterSpacing