+++ Sources/VComponents/Components/Item Pickers/Segmented Picker/VSegmentedPickerUIModel.swift
@@ -62,11 +62,11 @@ public struct VSegmentedPickerUIModel {
         /// Row content margin. Defaults to `2`.
         public var contentMargin: CGFloat = 2
         
-        /// Header line limit. Defaults to `1`.
-        public var headerLineLimit: Int? = 1
+        /// Header title type. Defaults to `singleline`.
+        public var headerTitleType: VTextType = .singleLine
         
-        /// Footer line limit. Defaults to `5`.
-        public var footerLineLimit: Int? = 5
+        /// Footer title type. Defaults to `multiline` of `1...5` lines.
+        public var footerTitleType: VTextType = .multiLine(alignment: .leading, lineLimit: 1...5)
         
         /// Spacing between header, picker, and footer. Defaults to `3`.
         public var headerPickerFooterSpacing: CGFloat = 3